move sources to src/
